Allow --watch without accountKey value
Fixes a bug with `cli wallet create` where if no value is specified for the --watch option, the accountKey is assigned a value of "true" which is invalid and causes errors downstream. This fix leaves accountKey as null, and the resulting functionality appears to match wallet creation via HTTP request where `watchOnly` is true and `accountKey` is not given a value.