Replace added() with test() in hasTX() definition
The function Pool.prototype.hasTX(hash), when called from a pool
with no mempool, checks if the hash is in this.txFilter. For this
check the txFilter method added() was previously used, but this
had the side effect of adding the hash to the txFilter. To avoid
this, the txFilter method test() is used instead with the exact
same inputs.