wallet: add "accountPath" to all responses dealing with "accountIndex"
Motivation:
In case of watch-only wallets, accountIndex is "arbitrary".
It simply maps to internal walletdb account instead of
BIP44 account path as defined in original accountKey.

It is important to return the actual accountIndex for use-cases
such as hardware-wallets, so clients can easily reproduce the
original path for this account.

To maintain backwards compatbility with existing wallets,
it is easiest to add this as a separate accountPath field
instead of refactoring all the existing uses of accountIndex.