test: browser testing
This commit alters the npm script
`test-browser` so that it runs bmocha
with the -H flag, which uses bpkg to
bundle the code and runs it in the
default headless browser. The previous
functionality of the `test-browser`
script was moved to `test-js`, which
runs the tests with the JavaScript
bcrypto backend. Note that this requires
bpkg to be installed globally.