script, tx, workers: Script.verify() accepts array of all input coins
Taproot introduces a sighash algorithm that requires a serialization
of the values of ALL the input coins into a transaction. Currently,
all the TX check/verify and Script verify methods expect a single
coin to be passed in, to check one specific input at a time. This
commit updates the signature for several functions throughout the
codebase to pass an array of ALL input coins when verifying a TX
input. Legacy behavior is preserved by using the index argument
to pick out the one coin being verified.

Many of the arguments expected by Script.verify() can be found
redundantly in either the tx object or the coins array it is passed.
Therefore the arguments to this function can be greatly simplified:

OLD:
verify(input, witness, output, tx, index, value, flags)

NEW:
verify(tx, index, coins, flags)

This refactor also modifies the worker packets that handle input
checking (async signing will need to be updated as well in a future
commit). All regression tests, bench tests, and script fuzz tests
were checked and pass.