script: verifyProgram() accepts array of all input coins
Like its caller verify(), this function will need the values of
all the input coins to compute the signature hash for Taproot spends.

Unlike verify(), we don't pull the "output" script from the coin of
the input being checked, because that script may have actually come
from a P2SH redeem script (in the input itself).