Added unit tests for coin,js
Line 52:
Each assertion here will throw an error if input type is not valid. For example, version must be a number

Line 61:
Here we are calculating number of confirmations for a transaction output based on current block height. In the case used here, if the coin height is 100 and current block height is 99, the depth should be 0. If the current block height is 101, the depth should be 2.

Line 77: 
Here firstly a coin object is created with all the valid attributes such as version, height etc. It then converts the coin object to JSON format using .toJSON() and then creates new coin object by converting JSON back using .fromJSON. This makes sure serialization and deserialization functions are working properly.

Line 95:
Here we are creating a new coin object using the input object ‘options’. Here assert statements are checking if coin objects are having correct values for each property.